Allow operations with decimal numbers

+++ Calculator/Program.cs
@@ -3,7 +3,7 @@
     private static void Main(string[] args)
     {
         // Declare variables and then initialize to zero.
-        int num1 = 0; int num2 = 0; int result = 0;
+        double num1 = 0; double num2 = 0; double result = 0;
 
         // Display title as the C# console calculator app.
         Console.WriteLine("------------------------\r");
@@ -15,11 +15,11 @@
 
         // Ask the user to type the first number.
         Console.WriteLine("Type a number, and then press Enter");
-        num1 = Convert.ToInt32(Console.ReadLine());
+        num1 = Convert.ToDouble(Console.ReadLine());
 
         // Ask the user to type the second number.
         Console.WriteLine("Type another number, and then press Enter");
-        num2 = Convert.ToInt32(Console.ReadLine());
+        num2 = Convert.ToDouble(Console.ReadLine());
 
         // Ask the user to choose an option.
         Console.WriteLine("Choose an option from the following list:");
